Creating our first custom hook useEntryDetails
