+++ src/components/NewEntryForm.js
@@ -1,32 +1,19 @@
-import { useState } from 'react';
 import { Form } from 'semantic-ui-react';
 import EntryForm from './EntryForm';
 import ButtonSaveOrCancel from './ButtonSaveOrCancel';
-import { useDispatch } from 'react-redux';
-import { addEntryRedux } from '../actions/entries.actions';
-import {v4 as uuidv4} from 'uuid';
+import useEntryDetails from '../hooks/useEntryDetails';
 
 function NewEntryForm() {
 
-  const [description, setDescription] = useState('');
-  const [value, setValue] = useState('');
-  const [isExpense, setIsExpense] = useState(true);
-  const dispatch = useDispatch();
-  
-  const addEntry = () => {
-    dispatch(
-      addEntryRedux({
-        id: uuidv4(),
-        description,
-        value,
-        isExpense
-      })
-    );
-
-    setDescription('');
-    setValue('');
-    setIsExpense(true);
-  }
+  const {
+    description,
+    setDescription,
+    value,
+    setValue,
+    isExpense,
+    setIsExpense,
+    addEntry
+  } = useEntryDetails();
 
     return (
         <Form unstackable>
